Add files via upload
Supplementary Material
The data folder contains all the datasets used in the experiment. Among them, 768_1.csv is the standardized dataset for Pima Indian diabetes, LMCH_.csv is the standardized dataset for Iraqi diabetes, and 2011-2020_1 is the standardized dataset from the National Health and Nutrition Examination Survey (NHANES).
All codes involved in the experiment have been implemented. The experimental codes are located in the "test" folder, and the other folders contain tool codes.